requirements.txt
gradio
numpy
tensorflow
keras-tuner
biopython
